etldocker/Flask App/requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.32
flasgger == 0.9.5
pandas == 1.3.5